requirements.txt
factorio-rcon-py==2.1.3
prometheus_client==0.21.1
python-dotenv==1.0.1
